move v2 and v1.2 only tests to their respective folders
